Make test suite pass on Mac OS
The current test suite make assumptions that are valid only on Linux. In
order to run the test suite on multiple platforms (e.g., Mac OS),
changes have been made to remedy the current behaviour.

- Don't assume `bash` is located in `/bin/bash`, let `/usr/bin/env`
decide where it's located.
- The `-P` flag on `grep` is not available in the BSD version. The `-E`
flag is available in both GNU Grep and the BSD version.
- The BSD version of `paste` does not implicitly use standard input if
no files are provided, explicitly tell `paste` to use standard input
using `-`.
- `find` does not have the `-executable` flag in the BSD version so we
have to fallback to using `-perm +111`. It's not equivalent, but does
the trick.
- `readlink` does not use the `-f` flag for the same purpose in the BSD
version as in the GNU version. In order to avoid problems, we're falling
back to a less elegant solution working across platforms.