switch to embedded version of termbox2.
consider this an unstable commit. termbox2 is a
line-by-line rewrite of termbox with some extra features,
more flexibility, stricter error handling, and a more
consistent api. it's also redesigned as a single-file
header library. this makes it easy to copy-paste into
other projects, which is what i've done here. at the time
of writing, the canonical termbox2 repo is here:

https://github.com/termbox/termbox2