fix bug with kmap fallthru for multi-input key bindings.
fallthru was not working properly for overlapping multi-input key
bindings. for example, if the current kmap (with fallthru allowed)
contained a key binding for `a x`, and its parent contained a key
binding for `a y`, then inputting `a y` would not execute the binding
in the parent kmap.

an additional issue arises here in that if the binding for `a y` is
`cmd_insert_data` then both inputs should count as input data, not
just the tail. (usually the inputs leading up to a multi-input
binding are simply consumed and forgotten.) this requires some
gymnastics. it would be a good exercise to see if we can
consolidate `pastebuf` and `input_trail`.