apply isearch_rule separately from normal srules
commit 7231152 broke isearch highlighting because if an earlier rule
matched, the isearch rule would not. prepending the rule would not
work either as it would potentially prevent other srules from applying
normally. for example, if we had an srule for "foobar" and currently
isearching "foo", we'd either apply whatever srule appeared first --
the "foobar" srule or the isearch rule -- but not both. therefore,
similar to range_srules, we have to apply isearch rules in a separate
pass.